Add firebase  (#590)
* Add firebase dependencies

* Initialize Firebase

* Add analytics store (need to figure out how to handle multiple analytics)
* Add logScreenView
* Add logging of firstLaunch

* Add dummy firebase-config.js for linting

* Added logAudioPlay and logAudioDuration functions

* Fixed lint error

* Add logShareContent

* Add logShareApp

* Add version tag

* Add dynamic imports of firebase

* Fix lint

* Change import config

* Check config for firebase user accounts

* Update logAudioPlay function

* Add 'export' keyword to convertFirebase

* Fix prettier and remove dead code

* Move logAudioPlay and logAudioDuration

* Add Analytics Parameters parsing

* Add null firebase configuration

* Log history item added

* Update first launch logging

* Rename ab_audio_play

* Fix lint error

---------

+++ convert/convertFirebase.ts
@@ -1,4 +1,4 @@
-import { existsSync, readFileSync, writeFileSync } from 'fs';
+import { existsSync, copyFile, unlinkSync, readFileSync, writeFileSync } from 'fs';
 import path from 'path';
 import { TaskOutput, Task } from './Task';
 
@@ -13,9 +13,14 @@ export function convertFirebase(dataDir: string, verbose: number) {
     const srcExists = existsSync(srcFile);
     const dstFile = path.join('src', 'lib', 'data', 'firebase-config.js');
     if (verbose) console.log(`FirebaseConfig: path=${srcFile}, exists=${srcExists}`);
-    const firebaseConfig =
-        'export ' + (srcExists ? readFileSync(srcFile) : 'const firebaseConfig = null;');
-    writeFileSync(dstFile, firebaseConfig);
+    if (srcExists) {
+        let content = readFileSync(srcFile, 'utf-8');
+        content = content.replace('const firebaseConfig', 'export const firebaseConfig');
+        writeFileSync(dstFile, content, 'utf-8');
+    } else {
+        const firebaseConfig = 'export const firebaseConfig = null;';
+        writeFileSync(dstFile, firebaseConfig);
+    }
 }
 
 export class ConvertFirebase extends Task {
